Fix 32-bit
Changes hardcoded Int64 -> Int. This makes the methods dispatch appropriately on 32-bit. For future reference @elevien , you usually shouldn't hardcode Int64 for this reason. No big deal though.

+++ src/coupling.jl
@@ -1,4 +1,4 @@
-SplitCoupledJumpProblem(prob::AbstractJumpProblem,prob_control::AbstractJumpProblem,aggregator::AbstractAggregatorAlgorithm,coupling_map::Vector{Tuple{Int64,Int64}})= JumpProblem(cat_problems(prob.prob,prob_control.prob),aggregator,build_split_jumps(prob,prob_control,coupling_map)...)
+SplitCoupledJumpProblem(prob::AbstractJumpProblem,prob_control::AbstractJumpProblem,aggregator::AbstractAggregatorAlgorithm,coupling_map::Vector{Tuple{Int,Int}})= JumpProblem(cat_problems(prob.prob,prob_control.prob),aggregator,build_split_jumps(prob,prob_control,coupling_map)...)
 
 # make new problem by joining initial_data
 function cat_problems(prob::DiscreteProblem,prob_control::DiscreteProblem)
@@ -85,7 +85,7 @@ cat_problems(prob_control::AbstractODEProblem,prob::AbstractSDEProblem) = cat_pr
 
 
 # this only depends on the jumps in prob, not prob.prob
-function build_split_jumps(prob::AbstractJumpProblem,prob_control::AbstractJumpProblem,coupling_map::Vector{Tuple{Int64,Int64}})
+function build_split_jumps(prob::AbstractJumpProblem,prob_control::AbstractJumpProblem,coupling_map::Vector{Tuple{Int,Int}})
   num_jumps = length(prob.discrete_jump_aggregation.rates)
   num_jumps_control = length(prob_control.discrete_jump_aggregation.rates)
   jumps = []
